docs: online

+++ docs.js
@@ -18,20 +18,52 @@ group:
 <API src="../src/<%= name %>.tsx"></API>
 `;
 
+const tpl1 =
+  `---
+title: <%= name %>
+order: <%= order %>
+group:
+  title: <%= groupTitle %>
+  order: <%= groupOrder %>
+  path: <%= groupPath %>
+---
+
+
+\`\`\`jsx
+
+
+
+// <%= title %>
+
+
+
+\`\`\`
+`;
+
 data.map((group, idx) => {
-  group.comps.map((item, subIdx) => {
-    item.order = subIdx;
-    item.groupOrder = idx;
-    item.groupPath = group.path;
-    item.groupTitle = group.title;
-    item.mobile = typeof item.mobile === 'boolean' ? item.mobile : true;
-
-    const fileName = path.resolve(__dirname, `./mdx/${item.name}.md`);
-
-    if (fs.existsSync(fileName)) {
-      fs.unlinkSync(fileName);
-    }
-
-    fs.writeFileSync(path.resolve(__dirname, `./mdx/${item.name}.md`), ejs.render(tpl, item));
-  });
+  handle(group, idx, Array.isArray(group.comps) ? 'comps' : 'list');
 });
+
+function handle(group, idx, listProp = 'comps') {
+  const list = group[listProp];
+  const isComps = listProp === 'comps';
+  if (Array.isArray(list))
+    list.map((item, subIdx) => {
+      item.order = subIdx;
+      item.groupOrder = idx;
+      item.groupPath = group.path;
+      item.groupTitle = group.title;
+      item.mobile = typeof item.mobile === 'boolean' ? item.mobile : true;
+
+      const fileName = path.resolve(__dirname, `./mdx/${item.name}.md`);
+
+      if (fs.existsSync(fileName)) {
+        fs.unlinkSync(fileName);
+      }
+
+      fs.writeFileSync(
+        path.resolve(__dirname, `./mdx/${item.name}.md`),
+        ejs.render(isComps ? tpl : tpl1, item)
+      );
+    });
+}
